requirements.txt
Flask==3.0.0
bcrypt==4.0.1 
pyjwt==2.8.0
python_dotenv==1.0.0
flask-cors==4.0.0